Make sure api deployment is separated out
