+++ golem-worker-service-base/src/gateway_api_deployment/mod.rs
@@ -1,0 +1,2 @@
+pub use http::*;
+mod http;
